MqttClient/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
paho-mqtt>=1.6.1
python-dotenv>=1.0.0
requests>=2.31.0